APS-475-V2: dataloaders and modules improvised

+++ retail/middleware/entitlement.service/src/app/v1/leave/leaves.resolver.ts
@@ -3,57 +3,68 @@ import {
   Query,
   Mutation,
   Args,
-  Info
 } from "@nestjs/graphql";
-import { graphqlKeys } from "@common/utilities";
 import { KeyValInput } from "@common/inputs/key-val.input";
 import {Leave} from "@app/v1/leave/leave.model";
 import {LeavesService} from "@app/v1/leave/leaves.service";
 import {LeaveInput} from "@app/v1/leave/leave.dto";
+import {HttpException, HttpStatus} from '@nestjs/common';
+import {MESSAGES} from '@common/constants';
+import {Fields} from '@common/decorators';
 
 @Resolver(Leave)
 export class LeavesResolver {
   constructor(private readonly leavesService: LeavesService) {}
 
   @Query(() => [Leave])
-  async leavesList(@Info() info): Promise<Leave[]> {
-    const keys = graphqlKeys(info);
-    return this.leavesService.list(keys);
+  async leavesList(@Fields() columns: string[]): Promise<Leave[]> {
+    return this.leavesService.list(columns);
   }
 
   @Query(() => Leave)
-  async findLeaveById(@Args('id') id: string, @Info() info): Promise<Leave> {
-    const keys = graphqlKeys(info);
-    return this.leavesService.findById(id, keys);
+  async findLeaveById(@Args('id') id: string, @Fields() columns: string[]): Promise<Leave> {
+    const leave: Leave = await this.leavesService.findById(id,columns);
+    if(!leave) throw new HttpException({
+      status: HttpStatus.NOT_FOUND,
+      error: MESSAGES.NOT_FOUND,
+    }, HttpStatus.NOT_FOUND);
+    return leave;
   }
 
   @Query(() => [Leave])
   async findLeaveBy(
       @Args('checks', { type: () => [KeyValInput] }) checks: KeyValInput[],
-      @Info() info
+      @Fields() columns: string[]
   ): Promise<Leave[]> {
-    const keys = graphqlKeys(info);
-    return this.leavesService.findByProperty(checks, keys);
+    return this.leavesService.findByProperty(checks, columns);
   }
 
   @Mutation(() => Leave)
-  async addLeave(@Args('input') input: LeaveInput, @Info() info): Promise<Leave> {
-    const keys = graphqlKeys(info);
-    return this.leavesService.create(input, keys);
+  async addLeave(@Args('input') input: LeaveInput, @Fields() columns: string[]): Promise<Leave> {
+    return this.leavesService.create(input, columns);
   }
 
   @Mutation(() => Leave)
   async updateLeave(
     @Args('id') id: string,
     @Args('input') input: LeaveInput,
-    @Info() info
+    @Fields() columns: string[]
   ): Promise<Leave> {
-    const keys = graphqlKeys(info);
-    return this.leavesService.update(id, input, keys);
+    const leave: Leave = await this.leavesService.findById(id,['id']);
+    if(!leave) throw new HttpException({
+      status: HttpStatus.NOT_FOUND,
+      error: MESSAGES.NOT_FOUND,
+    }, HttpStatus.NOT_FOUND);
+    return this.leavesService.update(id, input, columns);
   }
 
   @Mutation(() => Boolean)
   async deleteLeave(@Args('id') id: string): Promise<boolean> {
+    const leave: Leave = await this.leavesService.findById(id,['id']);
+    if(!leave) throw new HttpException({
+      status: HttpStatus.NOT_FOUND,
+      error: MESSAGES.NOT_FOUND,
+    }, HttpStatus.NOT_FOUND);
     return this.leavesService.delete(id);
   }
 }
