+++ retail/middleware/entitlement.service/src/app/v1/modules/module.resolver.ts
@@ -9,6 +9,9 @@ import {ModuleInput} from "@app/v1/modules/module.dto";
 import { KeyValInput } from "@common/inputs/key-val.input";
 import {Permission} from "@app/v1/permissions/permission.model";
 import {Fields} from '@common/decorators';
+import {User} from '@app/v1/users/user.model';
+import {HttpException, HttpStatus} from '@nestjs/common';
+import {MESSAGES} from '@common/constants';
 
 @Resolver(Module)
 export class ModuleResolver {
@@ -20,8 +23,13 @@ export class ModuleResolver {
   }
 
   @Query(() => Module)
-  async findModule(@Args('id') id: string, @Fields() columns: string[]): Promise<Module> {
-    return this.moduleService.findById(id, columns);
+  async findModuleById(@Args('id') id: string, @Fields() columns: string[]): Promise<Module> {
+    const module: Module = await this.moduleService.findById(id,columns);
+    if(!module) throw new HttpException({
+      status: HttpStatus.NOT_FOUND,
+      error: MESSAGES.NOT_FOUND,
+    }, HttpStatus.NOT_FOUND);
+    return module;
   }
 
   @Query(() => [Module])
@@ -43,11 +51,21 @@ export class ModuleResolver {
     @Args('input') input: ModuleInput,
     @Fields() columns: string[]
   ): Promise<Module> {
+    const module: Module = await this.moduleService.findById(id,['id']);
+    if(!module) throw new HttpException({
+      status: HttpStatus.NOT_FOUND,
+      error: MESSAGES.NOT_FOUND,
+    }, HttpStatus.NOT_FOUND);
     return this.moduleService.update(id, input, columns);
   }
 
   @Mutation(() => Boolean)
   async deleteModule(@Args('id') id: string): Promise<boolean> {
+    const module: Module = await this.moduleService.findById(id,['id']);
+    if(!module) throw new HttpException({
+      status: HttpStatus.NOT_FOUND,
+      error: MESSAGES.NOT_FOUND,
+    }, HttpStatus.NOT_FOUND);
     return this.moduleService.delete(id);
   }
